Feature/#46 post

+++ src/main/java/vanilla/witchtrial/service/PostServiceImpl.java
@@ -1,24 +1,30 @@
 package vanilla.witchtrial.service;
 
-import io.micrometer.common.util.StringUtils;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import vanilla.witchtrial.domain.Post;
+import vanilla.witchtrial.domain.UserAccount;
 import vanilla.witchtrial.dto.BoardDto;
 import vanilla.witchtrial.dto.PostDto;
+import vanilla.witchtrial.dto.UserPrincipal;
 import vanilla.witchtrial.global.common.constants.ErrorCode;
 import vanilla.witchtrial.global.exception.NotFoundException;
 import vanilla.witchtrial.repository.PostRepository;
+import vanilla.witchtrial.repository.UserRepository;
 
+@Slf4j
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Service
 public class PostServiceImpl implements PostService {
 
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
     @Override
     public Page<BoardDto.Response> getBoardList(BoardDto.Request request, Pageable pageable) {
@@ -41,31 +47,35 @@ public class PostServiceImpl implements PostService {
     @Transactional
     @Override
     public PostDto.Response saveNewPost(PostDto.Request postDto) {
-        Post post = PostDto.Request.toEntity(postDto);
+        UserAccount user = userRepository.getReferenceById(postDto.getUserPrincipal().id());
+        Post post = PostDto.Request.toEntity(postDto, user);
         postRepository.save(post);
-
         return PostDto.Response.from(post);
     }
 
     @Transactional
     @Override
-    public PostDto.Response updatePost(PostDto.UpdateRequest postDto) {
-        Post post = postRepository.findByIdWithDsl(postDto.getPostId()).orElseThrow(() -> new NotFoundException(ErrorCode.POST_NOT_FOUND));
-        if(!StringUtils.isBlank(postDto.getTitle())) {
+    public PostDto.Response updatePost(PostDto.UpdateRequest postDto) throws EntityNotFoundException, IllegalAccessException {
+        Post post = postRepository.getReferenceById(postDto.getPostId());
+        UserPrincipal userPrincipal = postDto.getUserPrincipal();
+
+        if(post.getUser().getId().equals(userPrincipal.id())) {
             post.setTitle(postDto.getTitle());
-        }
-        if(!StringUtils.isBlank(postDto.getContent())) {
             post.setContent(postDto.getContent());
+            return PostDto.Response.from(post);
+        } else {
+            throw new IllegalAccessException(ErrorCode.ILLEGAL_CLIENT_REQUEST.getMessage());
         }
-//        post.setHashtag(postDto.getHashtag());
-
-        return PostDto.Response.from(post);
     }
 
     @Transactional
     @Override
-    public void deletePost(Long id) {
-        Post post = postRepository.findByIdWithDsl(id).orElseThrow(() -> new NotFoundException(ErrorCode.POST_NOT_FOUND));
-        postRepository.delete(post);
+    public void deletePost(Long id, UserPrincipal userPrincipal) throws IllegalAccessException {
+        Post post = postRepository.getReferenceById(id);
+        if (post.getUser().getId().equals(userPrincipal.id())) {
+            postRepository.delete(post);
+        } else {
+            throw new IllegalAccessException(ErrorCode.ILLEGAL_CLIENT_REQUEST.getMessage());
+        }
     }
 }
